Renaming Util to Utility, because of a bug in Android Studio. This was producing false errors which did not stop the compilation.
