+++ app/src/main/java/de/seemoo/at_tracking_detection/database/models/device/types/Chipolo.kt
@@ -7,7 +7,7 @@ import androidx.annotation.DrawableRes
 import de.seemoo.at_tracking_detection.ATTrackingDetectionApplication
 import de.seemoo.at_tracking_detection.R
 import de.seemoo.at_tracking_detection.database.models.device.*
-import de.seemoo.at_tracking_detection.util.Util.getBitsFromByte
+import de.seemoo.at_tracking_detection.util.Utility.getBitsFromByte
 import timber.log.Timber
 
 class Chipolo(val id: Int) : Device() {
