+++ app/src/main/java/de/seemoo/at_tracking_detection/ui/dashboard/DashboardFragment.kt
@@ -1,11 +1,5 @@
 package de.seemoo.at_tracking_detection.ui.dashboard
 
-import android.Manifest
-import android.app.AlertDialog
-import android.bluetooth.BluetoothAdapter
-import android.bluetooth.BluetoothManager
-import android.content.Context
-import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -13,15 +7,9 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.navigation.NavDirections
-import androidx.navigation.fragment.findNavController
-import com.google.android.material.card.MaterialCardView
-import com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
 import dagger.hilt.android.AndroidEntryPoint
-import de.seemoo.at_tracking_detection.ATTrackingDetectionApplication
 import de.seemoo.at_tracking_detection.R
 import de.seemoo.at_tracking_detection.databinding.FragmentDashboardBinding
-import de.seemoo.at_tracking_detection.util.Util
 import timber.log.Timber
 import java.time.LocalDateTime
 import java.time.ZoneOffset
