+++ app/src/main/java/de/seemoo/at_tracking_detection/ui/devices/IgnoredDevicesFragment.kt
@@ -6,12 +6,12 @@ import androidx.navigation.fragment.FragmentNavigatorExtras
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.card.MaterialCardView
 import de.seemoo.at_tracking_detection.database.models.device.BaseDevice
-import de.seemoo.at_tracking_detection.util.Util
+import de.seemoo.at_tracking_detection.util.Utility
 
 class IgnoredDevicesFragment: DevicesFragment(showDevicesFound = false, showAllDevices = true ) {
     override val deviceItemListener: DeviceAdapter.OnClickListener
         get() = DeviceAdapter.OnClickListener { baseDevice: BaseDevice, materialCardView: MaterialCardView ->
-            if (!Util.checkAndRequestPermission(Manifest.permission.ACCESS_FINE_LOCATION)) {
+            if (!Utility.checkAndRequestPermission(Manifest.permission.ACCESS_FINE_LOCATION)) {
                 return@OnClickListener
             }
             val directions: NavDirections =
